check_list/requirements.txt
decorator==4.0.10
Django==1.10.3
django-extensions==1.7.4
django-filter==1.0.0
djangorestframework==3.5.3
Markdown==2.6.7
pathlib2==2.1.0
pexpect==4.2.1
psycopg2==2.6.2
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
Werkzeug==0.11.11
